Find commands with relative paths against the CWD.
which should return commands where the path is relative.  For example,
in a directory where a "foo" executable resides, executing "which ./foo"
should return "./foo".  I was treating relative paths as disallowed
completely, but they are just handled similarly to absolute paths - only
that singular location is checked and the PATH environment variable is
ignored.

This simplifies a lot and removes one more special rule that had to be
checked.