Extract class: PosixLocatorFactory.
As we try to add Windows support (for #5), we will need to have multiple
ways of handling the PATH environment variables.  This extracts that
behavior into a separate factory class in order to make it easily
extendable without hardcoding logic.  The code won't be
environment-specific and the user can specify which factory they would
like to use.

This also helps simplify the Locator by removing static methods from it
that can make things harder to test and reason about.