Add a platform factory that wraps the Posix/Windows factories.
This factory allows the user to not worry about whether they are on a
POSIX system or a Windows system (or some other systems in the future).
It determines which factory to use by a Windows-specific constant
PHP_WINDOWS_VERSION_BUILD that should only be defined on Windows
systems.

I'm unsure about how this will work on Cygwin systems that make some
things operate in a more POSIXy way.