Move path behaviors into new PathBuilder classes.
Because of differences in Windows and POSIX, there isn't a simple way to
use common code to handle both of them.  This moves the
_getPotentialCommandLocations logic into a new
PathBuilderInterface::getPermutations method that is defined for both
WindowsPathBuilder and PosixPathBuilder.

The interface to the locator factories does not change, so this does not
affect usage of them.  However, the Locator class itself has changed.
It no longer takes an array of paths in its constructor.  Instead it
takes an instance of PathBuilder, so in order to use it directly you
have to first create the appropriate PathBuilder.