Only support createFromEnvironment for all factories.  Fixes #8.
The issue I had with the PlatformLocatorFactory is that it's method
signature for the createFromPath method was different depending on
whether it was a Windows platform or Linux platform.  It also suggested
creating Windows locators without specifying PATHEXT information,
something I wouldn't recommend.

The getenv/habitat factory methods work fine - they properly abstract
out the information and provide a useful interface.

This also allows the abstract class to be converted to just an
interface, which is simpler to understand and work with externally.

I left the createFromPath methods in there for the WindowsLocatorFactory
and PosixLocatorFactory.  They could serve some use for somebody,
although directly constructing the PathBuilder's would probably be a
better approach.