Remove CLI command - it is nubs/which-cli now.
The CLI command was funny due to dependencies.  I didn't want to have
symfony/console be a dependency of the which library, but it is
necessary for the CLI command.  This meant that users wanting the
command would have to make sure to also install symfony/console
themselves.

As a separate package, this alleviates the concern.  I did a manual
subtree split plus cleanup in order to pull out the which command into
its own repository (without losing history :)).  People wanting the CLI
can pull that package in and it can have a hard dependency on
symfony/console.